Major refactor to facilitate adding more functionality.
I need to do wurtzite lattice constant calculation, which
has two lattice paramters and an internal degree of freedom.
My code as written cannot handle this case. After this refactor
it should be much easier to implement. This change breaks the
murnaghan capability for 2D materials (should be an easy fix)
and hasn't been tested with Abinit yet.